Rename class Float to Decimal to avoid PHP7 conflict

+++ src/Briedis/ApiBuilder/Items/Decimal.php
@@ -4,7 +4,7 @@
 namespace Briedis\ApiBuilder\Items;
 
 
-class Float extends BaseItem{
+class Decimal extends BaseItem{
 	const TYPE = 'float';
 
 	public function validateValue($value){
